Share SSA liveness for precise RegIR safepoints
